LastVersion

+++ src/main/java/com/example/sweater/service/AirplaneService.java
@@ -2,6 +2,7 @@ package com.example.sweater.service;
 
 import com.example.sweater.Models.AirplaneUpdateRM;
 import com.example.sweater.entity.Airplane_info;
+import com.example.sweater.entity.Airport;
 import com.example.sweater.repos.AirplaneRep;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,11 +15,15 @@ public class AirplaneService {
     @Autowired
     AirplaneRep airplaneRep;
 
-    public void save(Airplane_info airplane_info){
-        airplaneRep.save(airplane_info);
+    public Airplane_info save(Airplane_info airplane_info){
+        return airplaneRep.save(airplane_info);
     }
     public List<Airplane_info> getairplane_infos(){
         return airplaneRep.findAll();
+    }
+
+    public Airplane_info getAirplaneFromId(Long id){
+        return airplaneRep.findById(id).orElse(null);
     }
 
     public Airplane_info update(Long id, AirplaneUpdateRM airplaneUpdateRM) {
@@ -30,7 +35,9 @@ public class AirplaneService {
         return airplane_info;
     }
 
-    public void delete(Long id) {
-        airplaneRep.deleteById(id);
+    public Airplane_info delete(Long id) {
+        Airplane_info airplane_info=airplaneRep.findById(id).orElse(null);;
+        if(airplaneRep!=null) airplaneRep.deleteById(id);
+        return airplane_info;
     }
 }
