+++ src/main/java/com/example/sweater/service/AirportService.java
@@ -2,6 +2,7 @@ package com.example.sweater.service;
 
 import com.example.sweater.Models.AirportUpdateRM;
 import com.example.sweater.entity.Airport;
+import com.example.sweater.entity.Tickets_information;
 import com.example.sweater.repos.AirportRep;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,23 +15,28 @@ public class AirportService {
     @Autowired
     AirportRep airportRep;
 
-    public void save(Airport airport){
-        airportRep.save(airport);
+    public Airport save(Airport airport){
+       return airportRep.save(airport);
     }
     public List<Airport> getAirports(){
         return airportRep.findAll();
+    }
+
+    public Airport getAirportFromId(Long id){
+        return airportRep.findById(id).orElse(null);
     }
 
     public Airport update(Long id, AirportUpdateRM airportUpdateRM) {
 
         Airport airport =airportRep.findById(id).orElseThrow(EntityNotFoundException::new);
         airport.setCity(airportUpdateRM.getCity());
-
         airportRep.save(airport);
         return airport;
     }
 
-    public void delete(Long id) {
-        airportRep.deleteById(id);
+    public Airport delete(Long id) {
+        Airport airport=airportRep.findById(id).orElse(null);
+       if(airport!=null)  airportRep.deleteById(id);
+       return airport;
     }
 }
